requirements.txt
django>=3.2,<3.3
gunicorn
psycopg2-binary
python-dotenv
django-htmx
pint
django-storages
boto3
pillow
requests